pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "velloraq"
version = "0.1.0a0"
description = "Self-hosted serverless security posture management for AWS Lambda, Azure Functions, Google Cloud Functions, API gateways, IAM and storage."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Velloraq contributors" }]
keywords = ["serverless", "security", "aws", "azure", "gcp", "lambda", "cloud-functions", "devsecops"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Intended Audience :: Information Technology",
  "Intended Audience :: System Administrators",
  "Programming Language :: Python :: 3",
  "Topic :: Security",
]
dependencies = []

[project.urls]
Homepage = "https://github.com/velloraq/velloraq"
License = "https://opensource.org/license/mit"

[project.optional-dependencies]
saas = [
  "fastapi>=0.115",
  "uvicorn>=0.30",
  "SQLAlchemy>=2.0",
  "pg8000>=1.30",
  "bcrypt>=4.1",
  "PyJWT>=2.8",
]
aws = ["boto3>=1.34"]
azure = [
  "azure-identity>=1.16",
  "azure-mgmt-web>=7.2",
  "azure-mgmt-storage>=21.1",
  "azure-mgmt-authorization>=4.0",
  "azure-mgmt-resource>=23.0",
]
gcp = [
  "google-cloud-functions>=1.16",
  "google-cloud-storage>=2.16",
  "google-api-python-client>=2.120",
  "google-auth>=2.29",
]
dev = ["pytest>=8.0", "ruff>=0.4"]
all = [
  "fastapi>=0.115",
  "uvicorn>=0.30",
  "SQLAlchemy>=2.0",
  "pg8000>=1.30",
  "bcrypt>=4.1",
  "PyJWT>=2.8",
  "boto3>=1.34",
  "azure-identity>=1.16",
  "azure-mgmt-web>=7.2",
  "azure-mgmt-storage>=21.1",
  "azure-mgmt-authorization>=4.0",
  "azure-mgmt-resource>=23.0",
  "google-cloud-functions>=1.16",
  "google-cloud-storage>=2.16",
  "google-api-python-client>=2.120",
  "google-auth>=2.29",
]

[project.scripts]
velloraq = "velloraq.cli.entrypoint:main"
velloraq-worker = "velloraq.backend.workers.scan_worker:run_worker"
velloraq-migrate = "velloraq.backend.database.init_db:main"
velloraq-create-admin = "velloraq.backend.auth.create_admin:main"
slssec = "velloraq.cli.entrypoint:main"
slssec-worker = "velloraq.backend.workers.scan_worker:run_worker"
slssec-migrate = "velloraq.backend.database.init_db:main"
slssec-create-admin = "velloraq.backend.auth.create_admin:main"

[tool.setuptools.packages.find]
include = ["velloraq*", "app*", "serverless_security_scanner*"]

[tool.setuptools]
license-files = ["LICENSE"]

[tool.setuptools.package-data]
"velloraq.frontend" = ["*.html", "*.css", "*.js"]

[tool.ruff]
line-length = 100
target-version = "py310"
